03-automation-and-agend-based-systems/requirements.txt
crewai>=1.13.0
langchain>=1.2.15
langchain-openai>=1.1.12
python-dotenv
requests